+++ Custom/System - Install and upgrade/Paint.NET/Device_Require-Paint.NET_DoesNotRun.ps1
@@ -4,7 +4,7 @@
         Detects whether Paint.NET runs, returns $false if it does.
 
     .NOTES
-        Author:   Olav Rønnestad Birkeland
+        Author:   Olav Rønnestad Birkeland | github.com/o-l-a-v
         Created:  211205
         Modified: 211205